backend/requirements.txt
Flask
Flask-Cors
Flask-Mail
mysql-connector-python
python-dotenv